+++ ClassicCastbars/core/PoolManager.lua
@@ -51,6 +51,14 @@ function PoolManager:InitializeNewFrame(frame)
     frame.Text:ClearAllPoints()
     frame.Icon:SetPoint("LEFT", frame, -15, 0)
 
+    -- Dragonflight / retail
+    if WOW_PROJECT_ID == 1 then
+        frame.TextBorder:SetAlpha(0)
+        frame.BorderShield:SetTexture("Interface\\CastingBar\\UI-CastingBar-Small-Shield")
+        frame.Border:SetTexture("Interface\\CastingBar\\UI-CastingBar-Border-Small")
+        frame.Flash:SetTexture("Interface\\CastingBar\\UI-CastingBar-Flash-Small")
+    end
+
     -- Clear any scripts inherited from frame template
     frame:UnregisterAllEvents()
     frame:SetScript("OnLoad", nil)