+++ ClassicCastbars/CHANGELOG.md
@@ -1,6 +1,15 @@
 # ClassicCastbars
 
-## [v1.5.1](https://github.com/wardz/ClassicCastbars/tree/v1.5.1) (2022-08-16)
-[Full Changelog](https://github.com/wardz/ClassicCastbars/compare/v1.5.0...v1.5.1) [Previous Releases](https://github.com/wardz/ClassicCastbars/releases)
+## [v1.5.2](https://github.com/wardz/ClassicCastbars/tree/v1.5.2) (2022-11-19)
+[Full Changelog](https://github.com/wardz/ClassicCastbars/compare/v1.5.1...v1.5.2) [Previous Releases](https://github.com/wardz/ClassicCastbars/releases)
 
-- fix for player castbar testmode (#65)  
+- skip castID check here for channels  
+- dragonflight fixes  
+- bump toc  
+- revert to old texture in retail  
+- reanchor test castbar on target switch  
+- fix texture points for Dragonflight  
+- disable 'player' castbar in Dragonflight for now  
+- bump toc versions  
+- update libs  
+- v1 to v2